Server/requirements.txt
pypdf
fastapi
sqlalchemy
langchain_openai
langchain
pydantic
python-dotenv
langchain_community
faiss-cpu
pdfplumber
chromadb
google-generativeai
python-dotenv
langchain
PyPDF2
langchain_google_genai
uvicorn